Two test

+++ CardGame/UserControl.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using ConsoleTables;
 
@@ -124,10 +123,10 @@ namespace CardGame
                 string chosenAttribute;
 
                 Dictionary<string, double> dictioanryOfChoices = new Dictionary<string, double>();
-                dictioanryOfChoices["hp"] = player.topCard.HP / botPlayer.Hp; ;
-                dictioanryOfChoices["attack"] = player.topCard.Attack / botPlayer.Attack;
-                dictioanryOfChoices["defend"] = player.topCard.Defend / botPlayer.Defense;
-                dictioanryOfChoices["speed"] = player.topCard.Speed / botPlayer.Speed;
+                dictioanryOfChoices["hp"] = player.topCard.HP / (double)botPlayer.Hp; ;
+                dictioanryOfChoices["attack"] = player.topCard.Attack / (double)botPlayer.Attack;
+                dictioanryOfChoices["defend"] = player.topCard.Defend / (double)botPlayer.Defense;
+                dictioanryOfChoices["speed"] = player.topCard.Speed / (double)botPlayer.Speed;
 
                 chosenAttribute = dictioanryOfChoices.Aggregate((x, y) => x.Value > y.Value ? x : y).Key;
